Initial Commit
This is the initial commit with half the project done. I am submitting this because after this commit, I will begin adding the voting functionality, which is the main feature of this site. I would normally commit after all is done and clean up after...but I am doing this mainly for documentation purposes and also incase I totaly destroy everything, I have this commit to go back too.